Change TraceMe action to use metadatapopulator
