+++ pom.xml
@@ -155,7 +155,7 @@
       </dependency>
 -->
 
-      <!-- TraceMeAction.java dependency -->
+      <!-- TraceMe dependency -->
       <dependency>
          <groupId>javax.servlet</groupId>
          <artifactId>servlet-api</artifactId>
